Create README.md
# Text-Classification-using-Machine-Learning

Building a multiclass Text Classification Model from scratch without using the inbuilt 
Sklearn model in order to identify which among the twenty categories a particular document belongs. 
A lot of data preprocessing is also performed in this project . 
And finally the built model's performance is compared with the inbuilt sklearn model's performance.